+++ creep.setup.ranger.js
@@ -13,7 +13,7 @@ setup.maxCount = function(spawn){
     return _.filter(Game.flags, FLAG_COLOR.defense.filter).length;
 }
 setup.maxWeight = function(spawn){
-    return null; // no evaluation 
+    return 0; 
 }
 
 module.exports = setup;
